refactor: organize pages into feature-based folders
