+++ src/pages/Legal/Accessibility.tsx
