+++ src/pages/Legal/Disclaimer.tsx
