+++ src/pages/Legal/Privacy.tsx
